Implement search, as much as possible
There are a bunch of rough edges in this implementation.
The way the objects are serialized doesn't always make it
possible to deserialize them. Aggregations most notably
suffer from this.

I'm planning to try and NOT use the internal objects for
deserialization of aggs, but instead just parse the map
when you ask for something specific, but to use it, you'll
have to go through a transform on the client. Conversely,
if ES can start serializing more information, we can avoid
this problem entirely.

Closes #1
Related #5 #6